# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assignments/{assignmentId}/practice:
    post:
      operationId: generatePracticeAssignment
      tags: [practice]
      summary: Generate a fresh, never-graded practice version of an assignment (infinite)
      parameters:
        - name: assignmentId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeAssignmentSet"
  /practice/assignment-sessions/{sessionId}/grade:
    post:
      operationId: gradePracticeAssignment
      tags: [practice]
      summary: Grade all answers in a practice-assignment set with rich coaching feedback
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PracticeAssignmentGradeInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeAssignmentGrade"
...
components:
  schemas:
...
    KeystrokeTrace:
      type: object
      required: [keystrokeCount, eraseCount, durationMs]
      properties:
        keystrokeCount: { type: integer }
        eraseCount: { type: integer }
        bulkInsertCount: { type: integer, default: 0 }
        longestBulkInsertChars: { type: integer, default: 0 }
        rewriteSegments: { type: integer, default: 0 }
        durationMs: { type: integer }
...
    PracticeAssignmentSet:
      type: object
      required: [sessionId, assignmentId, title, kind, instructions, problems]
      properties:
        sessionId: { type: integer }
        assignmentId: { type: integer }
        title: { type: string }
        kind: { type: string }
        instructions: { type: ["string", "null"] }
        problems:
          type: array
          items:
            $ref: "#/components/schemas/PracticeAssignmentProblem"
    PracticeAssignmentProblem:
      type: object
      required: [id, prompt, topicId]
      properties:
        id: { type: integer }
        prompt: { type: string }
        topicId: { type: integer }
        topicTitle: { type: ["string", "null"] }
    PracticeAssignmentGradeInput:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items:
            $ref: "#/components/schemas/PracticeAssignmentAnswer"
    PracticeAssignmentAnswer:
      type: object
      required: [problemId, answer]
      properties:
        problemId: { type: integer }
        answer: { type: string }
        trace:
          $ref: "#/components/schemas/KeystrokeTrace"
    PracticeAssignmentGrade:
      type: object
      required: [score, total, percent, results]
      properties:
        score: { type: number }
        total: { type: number }
        percent: { type: number }
        results:
          type: array
          items:
            $ref: "#/components/schemas/PracticeAssignmentResult"
    PracticeAssignmentResult:
      type: object
      required: [problemId, prompt, correct, correctAnswer, feedback]
      properties:
        problemId: { type: integer }
        prompt: { type: string }
        correct: { type: boolean }
        userAnswer: { type: string }
        correctAnswer: { type: string }
        feedback: { type: string }
...